tools/fonts/requirements.txt
Pillow>=8.0.0
numpy
PyQt6